[#85] NSManagedObjectContext Extension 추가
